# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  aclmodels.Acl2Scope:
    enum:
    - ""
    - ror
    - cluster
    - project
    - datacenter
    type: string
    x-enum-comments:
      Acl2ScopeRor: ROR
      Acl2ScopeUnknown: unknown
    x-enum-varnames:
    - Acl2ScopeUnknown
    - Acl2ScopeRor
    - Acl2ScopeCluster
    - Acl2ScopeProject
    - Acl2ScopeDatacenter
...
  apiresourcecontracts.ResourceAction:
    enum:
    - Add
    - Delete
    - Update
    type: string
    x-enum-varnames:
    - K8sActionAdd
    - K8sActionDelete
    - K8sActionUpdate
...
  apiresourcecontracts.ResourceOwnerReference:
    properties:
      scope:
        allOf:
        - $ref: '#/definitions/aclmodels.Acl2Scope'
        description: cluster, workspace,...
      subject:
        description: ror id eg clusterId or workspaceName
        type: string
    type: object
  apiresourcecontracts.ResourceUpdateModel:
    properties:
      action:
        $ref: '#/definitions/apiresourcecontracts.ResourceAction'
      apiVersion:
        type: string
      hash:
        type: string
      kind:
        type: string
      owner:
        $ref: '#/definitions/apiresourcecontracts.ResourceOwnerReference'
      resource: {}
      uid:
        type: string
    type: object
...
info:
  contact:
    name: Privat Sky
    url: http://ror.sky.nhn.no
  description: 'ROR-API, need any help? Go to channel #drift-sdi-devops in norskhelsenett.slack.com
    slack workspace'
  title: Swagger ROR-API
  version: "0.1"
paths:
...
  /v1/resources/uid/{uid}:
    delete:
      consumes:
      - application/json
      description: Delete a resources
      parameters:
      - description: UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: bool
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Delete a resource by uid
      tags:
      - resources
    head:
      consumes:
      - application/json
      description: Get a list of cluster resources
      parameters:
      - description: The kind of the owner, currently only support 'Cluster'
        enum:
        - ""
        - ror
        - cluster
        - project
        - datacenter
        in: query
        name: ownerScope
        required: true
        type: string
      - description: The name og the owner
        in: query
        name: ownerSubject
        required: true
        type: string
      - description: UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Check cluster resource by uid
      tags:
      - resources
    put:
      consumes:
      - application/json
      description: Update a resources
      parameters:
      - description: UID
        in: path
        name: uid
        required: true
        type: string
      - description: ResourceUpdate
        in: body
        name: resourcereport
        required: true
        schema:
          $ref: '#/definitions/apiresourcecontracts.ResourceUpdateModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: bool
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Update resource by uid
      tags:
      - resources
...
  /v2/resources/uid/{uid}:
    delete:
      consumes:
      - application/json
      description: Delete a resources
      parameters:
      - description: UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: bool
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Delete a resource by uid
      tags:
      - resources
    head:
      consumes:
      - application/json
      description: Get a list of cluster resources
      parameters:
      - description: The kind of the owner, currently only support 'Cluster'
        enum:
        - ""
        - ror
        - cluster
        - project
        - datacenter
        in: query
        name: ownerScope
        required: true
        type: string
      - description: The name og the owner
        in: query
        name: ownerSubject
        required: true
        type: string
      - description: UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Check cluster resource by uid
      tags:
      - resources
    put:
      consumes:
      - application/json
      description: Update a resources
      parameters:
      - description: UID
        in: path
        name: uid
        required: true
        type: string
      - description: ResourceUpdate
        in: body
        name: resourcereport
        required: true
        schema:
          $ref: '#/definitions/apiresourcecontracts.ResourceUpdateModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: bool
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Update resource by uid
      tags:
      - resources
...
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
  ApiKey:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"